# Conflicts:
#	app/Http/Controllers/Account/PropertiesController.php
#	app/Property.php
#	resources/views/account/properties/form-video-thumb.blade.php
#	resources/views/account/properties/form.blade.php

+++ app/Http/Controllers/Account/PropertiesController.php
@@ -200,7 +200,7 @@ class PropertiesController extends \App\Http\Controllers\AccountController
 		exit;
 	}
 
-	public function create()
+	public function create($slug = null)
 	{
 		$modes = \App\Property::getModeOptions();
 		$types = \App\Property::getTypeOptions();
@@ -222,7 +222,26 @@ class PropertiesController extends \App\Http\Controllers\AccountController
 
 		$current_tab = session('current_tab', 'general');
 
-		return view('account.properties.create', compact('modes','types','energy_types','services','countries','states','cities','country_id','managers','current_tab', 'districts'));
+		if($slug) {
+			$property = $this->site->properties()->whereTranslation('slug', $slug)
+					->with([ 'catches' => function($query){
+							$query->with('buyer');
+					}])->first();
+
+			$property->ref = '';
+			$marketplaces = $this->site->marketplaces()
+							->wherePivot('marketplace_enabled','=',1)
+							->withSiteProperties($this->site->id)
+							->enabled()->orderBy('name')->get();
+
+
+			$states = \App\Models\Geography\State::enabled()->where('country_id', $property->country_id)->lists('name','id');
+			$cities = \App\Models\Geography\City::enabled()->where('state_id', $property->state_id)->lists('name','id');
+			$districts = $this->site->districts()->orderBy('name')->lists('name','id');
+		}
+
+		return view('account.properties.create', compact('modes','types','energy_types',
+				'services','countries','states','cities','country_id','managers','current_tab', 'districts', 'property', 'marketplaces'));
 	}
 
 	public function store()
@@ -291,6 +310,26 @@ class PropertiesController extends \App\Http\Controllers\AccountController
 			$data[$field] = $this->request->input($field);
 		}
 		$catch->update($data);
+
+		//New video
+		if($this->request->input('video_link')) {
+			$link = $this->request->input('video_link');
+			$saved = $this->saveVideoLink($property, $link);
+			if(!$saved) {
+				return redirect()->back()->withInput()
+					->with('error', trans('account/properties.video.error'));
+			}
+		}
+
+		//Duplicate videos
+		$duplicateVideoIds = $this->request->input('duplicate_videos', []);
+		foreach($duplicateVideoIds as $duplicateVideoId){
+			$video = Videos::where('id', $duplicateVideoId)->first();
+			if(!$video) {
+				continue;
+			}
+			$this->saveVideoLink($property, $video->link);
+		}
 
 		$property = $this->site->properties()->find($property->id);
 
@@ -393,6 +432,16 @@ class PropertiesController extends \App\Http\Controllers\AccountController
 			return redirect()->back()->withInput()->with('error', trans('general.messages.error'));
 		}
 
+		//New video
+		if($this->request->input('video_link')) {
+			$link = $this->request->input('video_link');
+			$saved = $this->saveVideoLink($property, $link);
+			if(!$saved) {
+				return redirect()->back()->withInput()
+					->with('error', trans('account/properties.video.error'));
+			}
+		}
+
 		if($isPriceFall){
 			if($this->site->alert_config === null ||
 					$this->site->alert_config['bajada']['agentes']){
@@ -425,6 +474,39 @@ class PropertiesController extends \App\Http\Controllers\AccountController
 		return redirect()->action('Account\PropertiesController@edit', $property->slug)->with('current_tab', $this->request->input('current_tab'))->with('success', trans('account/properties.saved'));
 	}
 
+	/**
+	 * @param Property $property
+	 * @param string $link
+	 * @return boolean|null
+	 */
+	private function saveVideoLink($property, $link){
+		$isVimeo = Videos::isVideoVimeo($link);
+		$isYoutube = Videos::isVideoYoutube($link);
+
+		if(!$isVimeo && !$isYoutube){
+			return false;
+		}
+
+		$remoteThumbnail = $isVimeo ? Videos::getVimeoThumbnail($link) :
+			Videos::getYoutubeThumbnail($link);
+
+		if($remoteThumbnail) {
+			$imageName = $isVimeo ? Videos::getVimeoCode($link) :
+				Videos::getYouTubeCode($link);
+
+			$localName = $property->video_path . '/' . $imageName . '.jpg';
+			$thumbUploaded = copy($remoteThumbnail, $localName);
+
+			if($thumbUploaded) {
+				$video = new Videos();
+				$video->property_id = $property->id;
+				$video->link = $link;
+				$video->thumbnail = $imageName . '.jpg';
+				return $video->save();
+			}
+		}
+	}
+
 	public function show($slug)
 	{
 		// Get property
@@ -447,7 +529,7 @@ class PropertiesController extends \App\Http\Controllers\AccountController
 		}
 
 		$current_tab = session('current_tab', old('current_tab', 'tab-general'));
-				
+
 		return view('account.properties.show', compact('property', 'current_tab'));
 	}
 
@@ -1017,6 +1099,26 @@ class PropertiesController extends \App\Http\Controllers\AccountController
 				{
 					$image = $property->images()->find($image_id);
 
+					if(!$image && $is_new) {//duplicate image
+						$imageToCopy = \App\Models\Property\Images::find($image_id);
+						$parentImagePath = $imageToCopy->property->image_path . '/' . $imageToCopy->image;
+
+						$newImageName = $property->id."_{$imageToCopy->image}";
+						$newImagepPath = $property->image_path . '/' . $newImageName;
+
+						if(file_exists($parentImagePath)) {
+							copy($parentImagePath, $newImagepPath);
+						}
+
+						//save image object even if copy failed. or file do not exist.
+						$image = $property->images()->create([
+							'image' => $newImageName,
+							'position' => $position,
+							'created_at' => new \DateTime(),
+							'updated_at' => new \DateTime(),
+						]);
+					}
+
 					$updateFields = [
 						'default' => 0,
 						'position' => $position,
@@ -1051,7 +1153,7 @@ class PropertiesController extends \App\Http\Controllers\AccountController
 		// Deleted images
 		foreach ($property->images as $image)
 		{
-			if ( !in_array($image->id, $preserve) )
+			if ( !in_array($image->id, $preserve) && !$is_new)
 			{
 				@unlink( public_path("sites/{$property->site_id}/properties/{$property->id}/{$image->image}") );
 				$image->delete();
@@ -1096,6 +1198,16 @@ class PropertiesController extends \App\Http\Controllers\AccountController
 		if ( $default_image )
 		{
 			$default_image->update([ 'default'=>1 ]);
+		}
+
+		$videoIds = $this->request->input('videos', []);
+
+		foreach($property->videos as $video) {
+			if(!in_array($video->id, $videoIds)){
+				$videoThumbPath = $property->video_path . '/' . $video->thumbnail;
+				@unlink( $videoThumbPath );
+				$video->delete();
+			}
 		}
 
 		return true;
